Makefile.in: Changed CPU_TYPE to ARCH to have a more consistent naming. Removed -I$(top_srcdir) from CPPFLAGS as it is not needed any more. Added $(builddir)/ to version_date.h. Added missing $(top_srcdir)/ to ChangeLog. Added DEPEND to use architecture dependent depend files. Commented out the makefile definiton with WORD and WORDS, as we now have architecture specific depend files. Instead now makefile is defined to $(builddir)/Makefile.
darwin/Makefile: Added targets depend{,_i386,_ppc} to generate dependencies and distclean{,_i386,_ppc} to have clean and distclean targets similar to the Makefile. Changed usage of make again to use the -C option. Thus now the targets are build in the same way then for other systems. This has been necessary because the builddir now is fixed to "." and can't be changed. On the other hand it's possible now, as we have the dependencies in other files.

configure.ac: Added the option -mmacosx-version-min to ISYSROOT, this should fix the build on Leopard without the crt1.o workaround. Changed the srcdir for darwin back to ".", as now make is used with the -C option.

INSTALL: Fixed the path description to the executable.

mfp.cpp: Commented out the panicbug "mfp_do_interrupt called when another higher priority interrupt is running", as it is issued continuously when MagiC runs.

ndebug.{cpp,h}: Added issavettyvalid and thus fixed ndebug in case the terminal is not supporting tcgetattr().

darwin/aranym.xcodeproj/project.pbxproj: Updated the Xcode project file to contain new files and to use SCM/CVS. Also cleaned up some settings.