Fix race condition in IKBD layer
MFP generates interrupt on transition for GPIP IKBD pin. Pin is set to
1 when IKBD buffer is empty, and pin is 0 when there is data from IKBD.
Force 0->1 then 1->0 transition when an interrupt must be generated,
in some cases, many bytes are waiting in IKBD layer, but the pin is
already set to 0 by first processed byte, and no more interrupts are
generated on MFP side.

Maybe modify the MFP emulator to ask IKBD which state the GPIP pin
should be ?