Do not use strncpy for copying long home folder strings
This fixes a compiler warning that GCC 9.1 emits:

In file included from /usr/include/string.h:494,
                 from /usr/include/SDL2/SDL_stdinc.h:60,
                 from /usr/include/SDL2/SDL_main.h:25,
                 from /usr/include/SDL2/SDL.h:32,
                 from ./include/SDL_compat.h:36,
                 from ./include/parameters.h:32,
                 from Unix/host_filesys_unix.cpp:23:
In function 'char* strncpy(char*, const char*, size_t)',
    inlined from 'static char* HostFilesys::getHomeFolder(char*, unsigned int)' at Unix/host_filesys_unix.cpp:41:10,
    inlined from 'static char* HostFilesys::getConfFolder(char*, unsigned int)' at Unix/host_filesys_unix.cpp:54:42:
/usr/include/bits/string_fortified.h:106:34: error: 'char* __builtin_strncpy(char*, const char*, long unsigned int)' specified bound 512 equals destination size [-Werror=stringop-truncation]
  106 |   return __builtin___strncpy_chk (__dest, __src, __len, __bos (__dest));

We must not use strncpy if $HOME is longer than the buffer, since this
results in a non-terminated string and certainly causes crashes or
other bad effects later. The only thing that we can do here is to
ignore the $HOME folder here and use the current directory instead.